Update Readinza app details and add IPA file.
This commit updates the `apps.json` file to reflect the new version "0.0.4" of the Readinza app.
The developer name is changed to "Readinza", the version date is updated, and the version description is now "Update release.".
The app size has also been updated.
Additionally, this commit includes the `Readinza.ipa` file.